implemented get_scenarios route using dataset
